Выносит одинаковую разметку в GameScreen, добавляет тест компонента GameScreen
