+++ src/components/guess-genre-screen/guess-genre-screen.jsx
@@ -1,12 +1,6 @@
 import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 
-const svgCircleStyles = {
-  filter: `url(#blur)`,
-  transform: `rotate(-90deg) scaleY(-1)`,
-  transformOrigin: `center`
-};
-
 class GuessGenreScreen extends PureComponent {
   constructor(props) {
     super(props);
@@ -32,55 +26,36 @@ class GuessGenreScreen extends PureComponent {
     const {genre, answers} = question;
 
     return (
-      <section className="game game--genre">
-        <header className="game__header">
-          <a className="game__back" href="#">
-            <span className="visually-hidden">Сыграть ещё раз</span>
-            <img className="game__logo" src="img/melody-logo-ginger.png" alt="Угадай мелодию"/>
-          </a>
+      <section className="game__screen">
+        <h2 className="game__title">Выберите {genre} треки</h2>
+        <form className="game__tracks" onSubmit={(evt) => {
+          evt.preventDefault();
+          onAnswer(question, this.state.answer);
+        }}>
 
-          <svg xmlns="http://www.w3.org/2000/svg" className="timer" viewBox="0 0 780 780">
-            <circle className="timer__line" cx="390" cy="390" r="370" style={svgCircleStyles}/>
-          </svg>
-
-          <div className="game__mistakes">
-            <div className="wrong"></div>
-            <div className="wrong"></div>
-            <div className="wrong"></div>
-          </div>
-        </header>
-
-        <section className="game__screen">
-          <h2 className="game__title">Выберите {genre} треки</h2>
-          <form className="game__tracks" onSubmit={(evt) => {
-            evt.preventDefault();
-            onAnswer(question, this.state.answer);
-          }}>
-
-            {answers.map((el) => {
-              return (
-                <div className="track" key={el.id}>
-                  <button className="track__button track__button--play" type="button"></button>
-                  <div className="track__status">
-                    <audio src={el.src}></audio>
-                  </div>
-                  <div className="game__answer">
-                    <input className="game__input visually-hidden" type="checkbox" name="answer" value={el.genre} id={`answer-${el.id}`} onChange={() => {
-                      const oldState = this.state.answer;
-                      oldState[el.id] = !oldState[el.id];
-                      const newState = oldState;
-                      this.setState({answers: newState});
-                    }
-                    }/>
-                    <label className="game__check" htmlFor={`answer-${el.id}`}>Отметить</label>
-                  </div>
+          {answers.map((el) => {
+            return (
+              <div className="track" key={el.id}>
+                <button className="track__button track__button--play" type="button"></button>
+                <div className="track__status">
+                  <audio src={el.src}></audio>
                 </div>
-              );
-            })}
+                <div className="game__answer">
+                  <input className="game__input visually-hidden" type="checkbox" name="answer" value={el.genre} id={`answer-${el.id}`} onChange={() => {
+                    const oldState = this.state.answer;
+                    oldState[el.id] = !oldState[el.id];
+                    const newState = oldState;
+                    this.setState({answers: newState});
+                  }
+                  }/>
+                  <label className="game__check" htmlFor={`answer-${el.id}`}>Отметить</label>
+                </div>
+              </div>
+            );
+          })}
 
-            <button className="game__submit button" type="submit">Ответить</button>
-          </form>
-        </section>
+          <button className="game__submit button" type="submit">Ответить</button>
+        </form>
       </section>
     );
   }
